Add files via upload
This one of the project i did during my second year at Mangosuthu University of Technology, i was introduced to C# programming language and i spend a lot of time practising. After few week of practising i then design and develop this project it runs succesfully. 

This is an #InventoryManagementSystem